+++ src/features/Treege/components/FieldsSelect/index.ts
@@ -1,0 +1,4 @@
+import FieldsSelect from "./FieldsSelect";
+
+export default FieldsSelect;
+export * from "./FieldsSelect";