+++ src/features/Treege/context/TreegeProvider.tsx
@@ -1,22 +1,26 @@
 import { ReactNode, useMemo, useReducer, useState } from "react";
 import { treeDefaultValue, TreegeContext } from "@/features/Treege/context/TreegeContext";
-import treeReducer from "@/features/Treege/reducer/treeReducer";
+import treeReducer, { setTree } from "@/features/Treege/reducer/treeReducer";
 import type { TreeNode } from "@/features/Treege/type/TreeNode";
+import useWorkflowQuery from "@/services/workflows/query/useWorkflowQuery";
 import { version } from "~/package.json";
 
 interface TreegeProviderProps {
   children: ReactNode;
   endPoint?: string;
   initialTree?: TreeNode;
+  initialTreeId?: string;
 }
 
-const TreegeProvider = ({ children, endPoint, initialTree }: TreegeProviderProps) => {
+const TreegeProvider = ({ children, endPoint, initialTree, initialTreeId }: TreegeProviderProps) => {
   const [currentHierarchyPointNode, setCurrentHierarchyPointNode] = useState(treeDefaultValue.currentHierarchyPointNode);
   const [modalOpen, setModalOpen] = useState(treeDefaultValue.modalOpen);
   const [treeModalOpen, setTreeModalOpen] = useState(treeDefaultValue.treeModalOpen);
   const [treePath, setTreePath] = useState(treeDefaultValue.treePath);
-  const [currentTree, setCurrentTree] = useState(treeDefaultValue.currentTree);
   const [tree, dispatchTree] = useReducer(treeReducer, initialTree || treeDefaultValue.tree);
+  const [currentTree, setCurrentTree] = useState(
+    initialTreeId ? { ...treeDefaultValue.currentTree, id: initialTreeId } : treeDefaultValue.currentTree
+  );
 
   const value = useMemo(
     () => ({
@@ -35,8 +39,17 @@ const TreegeProvider = ({ children, endPoint, initialTree }: TreegeProviderProps
       treePath,
       version,
     }),
-    [currentHierarchyPointNode, modalOpen, treeModalOpen, treePath, tree, endPoint, currentTree]
+    [currentHierarchyPointNode, modalOpen, treeModalOpen, treePath, tree, currentTree, endPoint]
   );
+
+  // Fetch initial tree
+  useWorkflowQuery(currentTree.id, {
+    enabled: !!initialTreeId,
+    onSuccess: (response) => {
+      setCurrentTree({ id: response?.id, name: response?.label });
+      dispatchTree(setTree(response.workflow));
+    },
+  });
 
   return <TreegeContext.Provider value={value}>{children}</TreegeContext.Provider>;
 };