+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # [Versions](https://github.com/Tracktor/treege/releases)
 
-## v0.4.20
-- Add repeatable option
+## v0.5.0
+- Add new props "initialTreeId"
+- Remove helper text name to tree name text field
+- Default tree name props
+- Fix translation
+- Improve tree select
+- Disable some features when endpoint is not provided
+- Documentation